Relax expressions in port definition : extract a forbidden-nodes target
