+++ src/io/inputs/CMakeLists.txt
@@ -1,5 +1,6 @@
 OMESSAGE("Antares IO-Inputs")
 
+add_subdirectory(forbidden-nodes)
 add_subdirectory(model-converter)
 add_subdirectory(yml-model)
 add_subdirectory(yml-optim-config)
