+++ src/io/inputs/forbidden-nodes/include/antares/io/inputs/forbidden-nodes/ForbiddenNodes.h
@@ -11,7 +11,7 @@
 #include "antares/expressions/nodes/FunctionNode.h"
 #include "antares/expressions/nodes/Node.h"
 
-namespace Antares::IO::Inputs::ModelConverter
+namespace Antares::IO::Inputs::ForbidNodes
 {
 
 template<typename NodeType>
@@ -77,4 +77,4 @@ private:
     std::map<std::type_index, std::set<std::type_index>> rules_; // Parent --> set of children
 };
 
-} // namespace Antares::IO::Inputs::ModelConverter
+} // namespace Antares::IO::Inputs::ForbidNodes
