+++ src/packaging/CMakeLists.txt
@@ -100,6 +100,7 @@ set(TARGET_LIBS #No alias
         modeler-optimisation-container
         expressions
         optim-model-filler
+		forbidden-nodes
         yml-system
         yml-model
         model-converter
